server/requirements.txt
altgraph==0.17.4
dill==0.4.0
fire==0.7.1
mpmath==1.3.0
numpy==2.3.3
packaging==25.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyYAML==6.0.3
setuptools==80.9.0
sympy==1.14.0
termcolor==3.1.0
